+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### New Features
 - Add Tampa Intl. (KTPA) [#647](https://github.com/openscope/openscope/issues/647)
 - Add Theodore Francis Green Memorial State Airport (Providence, US) [#107](https://github.com/openscope/openscope/issues/107)
+- New Airport: KDTW (Detroit Metropolitan Wayne County Airport, MI) [#559](https://github.com/openscope/openscope/issues/559)